Automating Python Libraries used into Excel sheet
This script will automate the task of finding which libraries are used in hundred and thousands of python files.  Allows to summarize all the libraries used in a bunch of python files into a single Excel sheet. This sheet could be later referred to find which libraries were used and how were they used.